Change - regen-repos checksum_of() implementation
There is a strong suspicion that regen-repos is triggering more often
than necessary.  For example, it's quite possible that packages are
being updated by mirrmiad in the external repositories even though none
of the inter-dependencies have actually changed.  While this commit
certainly does not achieve an ideal solution, it's predicted to be
significantly better.  Rather than triggering on any change within the
external repositories, it now looks only at the repo meta-data there.
It's acknowledged that this repo meta-data is changing even if the
inter-dependencies are not, but this should reduce the number of trigger
events caused by the regen-repos service performing its checks
concurrent with mirrmaid updating the mirrors.

This change brings another improvement in that only the content of this
meta-data can trigger a change.  Simply changing time-stamps, ownership,
etc. is no longer sufficient.