Change - smashd obsoletes mash-everything
This new event-driven tool (smashd) is now sufficiently well developed
to serve as a complete replacement for the scheduled/polling tool
(mash-everything).

While mash-everything got the job done, it was a horribly inefficient
design that caused lots of work in Koji and the filesystem in a manner
that was too often and not often enough at the same time.  Most runs
produced absolutely no difference in the package repositories because
there actually hadn't been any build or tag activity to justify the
work.  In other, less frequent cases, a build or tag event had just
occurred and the user would be waiting several minutes for the next
scheduled cycle to occur to have any observable effect.  Now that it's
event driven the latency only comes from two sources:

    - a quiescence period to ensure the run doesn't start so soon that
      it would fail to combine the entire group into a single, more
      efficient run

    - the actual execution time of the mash/sync process

In addition to the above gains, other improvements also are gained:

    - only those repositories that need by affected are updated

    - because of that, clients won't see repository meta-data changes
      unless there's actually changes in the repository contents meaning
      their operations will be more efficient which will further reduce
      load on the web server and file systems providing those
      repositories