Change - extend smashd to sign as needed
This eliminates our dependency on the sigulsign_unsigned.py that
I lifted from the Fedora Project's infrastructure team.  That served
it's purpose to get us going, but it was a rather big hammer.  With
smashd providing an event-driven process, things were improved but the
mashing process still had to be delayed to ensure that the RPM signing
process had time to occur.  That all changes here as the signing is now
part of the same event driven loop where RPM signing has been inserted
as just one additional step prior to mashing.