Bug - [smashd] don't sign empty build set
In a situation where one package is being promoted from one tag to
another, the first tag will have events, but no builds needing to be
signed.  This is a continuation of commit 00b1534 which classified
events as tag-in or tag-out, but failed to realize that the build set
could now be empty when signing.