Change - [smashd] only sign unsigned RPMs
Rather than blindly asking Sigul to sign all RPMs involved in the build,
the Signer class is now smart and asks Koji for a set of RPMs within the
tag that already signed.  Because KojiBuildInfo.rpms now returns a set,
simple set math makes it trivial to determine those not yet signed.