Bug - [gojira] connection errors are fatal
Some connection errors are naturally intermittent while others are
merely an administrative configuration error that will never succeed.
In any event, it's best if gojira logs a decent message and tries to
continue the best it can.  Perhaps it should distinguish between
temporary and permanent failures, but is unnecessarily complex at this
time.