Bug - [gojira] Thread exceptions may not be logged
It seems, at least with the syslog formatter, that any uncaught
exceptions occurring during BuildRootDependenciesMonitor.run() are not
logged unless the syslog is configured for the DEBUG level, which is not
the default.  This is bad behavior.  Such exceptions are now caught with
a basic "oh shit" handler and logged accordingly so as to not remain
hidden.