Bug - [gojira] Thread dies with incomplete metadata
It has been observed that the BuildRootDependenciesMonitor thread can
die unexpectedly in rare circumstances.  In the one observed case, the
HTTP HEAD request did not have the expected 'last-modified' field which
resulted in a KeyError and thread death.  This commit wrap this critical
code in a retry loop for some basic resilience.